+++ src/main/java/com/c9pay/storeservice/dto/product/ProductForm.java
@@ -1,4 +1,4 @@
-package com.c9pay.storeservice.dto;
+package com.c9pay.storeservice.dto.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
