+++ src/main/java/com/c9pay/storeservice/dto/sale/ProductSaleInfo.java
@@ -1,4 +1,4 @@
-package com.c9pay.storeservice.dto;
+package com.c9pay.storeservice.dto.sale;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProductPurchaseInfo {
+public class ProductSaleInfo {
     private long productId;
+    private String name;
+    private int price;
     private int amount;
 }
