+++ src/main/java/com/c9pay/storeservice/dto/sale/PurchaseInfo.java
@@ -1,5 +1,6 @@
-package com.c9pay.storeservice.dto;
+package com.c9pay.storeservice.dto.sale;
 
+import com.c9pay.storeservice.dto.QRInfo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
