+++ src/main/java/com/c9pay/storeservice/dto/store/StoreDetails.java
@@ -1,4 +1,4 @@
-package com.c9pay.storeservice.dto;
+package com.c9pay.storeservice.dto.store;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProductSaleInfo {
-    private long productId;
+public class StoreDetails {
+    private Long id;
     private String name;
-    private int price;
-    private int amount;
 }
